split dialogs into multiple modules
