+++ rnote-ui/po/POTFILES
@@ -23,7 +23,9 @@ rnote-ui/data/ui/penssidebar/shaperpage.ui
 rnote-ui/data/ui/penssidebar/toolspage.ui
 rnote-ui/data/ui/penssidebar/typewriterpage.ui
 
-rnote-ui/src/dialogs.rs
+rnote-ui/src/dialogs/mod.rs
+rnote-ui/src/dialogs/import.rs
+rnote-ui/src/dialogs/export.rs
 rnote-ui/src/appwindow/mod.rs
 rnote-ui/src/appwindow/imexport.rs
 rnote-ui/src/appwindow/appwindowactions.rs
