+++ rnote-ui/src/meson.build
@@ -68,10 +68,12 @@ rnote_ui_sources = files(
     'workspacebrowser/workspacelist.rs',
     'workspacebrowser/workspacelistentry.rs',
     'workspacebrowser/workspacerow.rs',
+    'dialogs/mod.rs',
+    'dialogs/import.rs',
+    'dialogs/export.rs',
     'appmenu.rs',
     'canvasmenu.rs',
     'config.rs',
-    'dialogs.rs',
     'globals.rs',
     'iconpicker.rs',
     'main.rs',
